+++ app/Http/Controllers/SummaryCategoryController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Category;
 use App\Option\Get;
 use App\Utilities\Header;
+use App\Utilities\RoutePermission;
+use App\Validators\Request\Route;
 use Illuminate\Http\JsonResponse;
 
 /**
@@ -19,11 +21,19 @@ class SummaryCategoryController extends Controller
     /**
      * Return a summary of the categories
      *
+     * @param $resource_type_id
+     *
      * @return JsonResponse
      */
-    public function index(): JsonResponse
+    public function index($resource_type_id): JsonResponse
     {
+        Route::resourceType(
+            $resource_type_id,
+            $this->permitted_resource_types
+        );
+
         $summary = (new Category())->totalCount(
+            $resource_type_id,
             $this->permitted_resource_types,
             $this->include_public
         );
@@ -45,13 +55,25 @@ class SummaryCategoryController extends Controller
     /**
      * Generate the OPTIONS request for the categories summary
      *
+     * @param $resource_type_id
+     *
      * @return JsonResponse
      */
-    public function optionsIndex(): JsonResponse
+    public function optionsIndex($resource_type_id): JsonResponse
     {
+        Route::resourceType(
+            $resource_type_id,
+            $this->permitted_resource_types
+        );
+
+        $permissions = RoutePermission::resourceType(
+            $resource_type_id,
+            $this->permitted_resource_types
+        );
+
         $get = Get::init()->
             setDescription('route-descriptions.summary_category_GET_index')->
-            setAuthenticationStatus(($this->user_id !== null) ? true : false)->
+            setAuthenticationStatus($permissions['view'])->
             option();
 
         return $this->optionsResponse($get, 200);